sounds n stuff

+++ code/modules/replicators/replicators_forcefields.dm
@@ -17,6 +17,9 @@
 	icon_state = "floor"
 
 /turf/simulated/floor/plating/airless/catwalk/forcefield/Destroy()
+	// to-do: sound
+	playsound(src, pick('sound/machines/arcade/gethit1.ogg', 'sound/machines/arcade/gethit2.ogg', 'sound/machines/arcade/-mana1.ogg', 'sound/machines/arcade/-mana2.ogg'), VOL_EFFECTS_MASTER, 80)
+
 	var/obj/structure/forcefield_node/FN = locate() in src
 	qdel(FN)
 
@@ -33,15 +36,16 @@
 
 /turf/simulated/floor/plating/airless/catwalk/forcefield/attackby(obj/item/C, mob/user)
 	if(istype(C, /obj/item/stack/tile) && !user.is_busy() && do_skilled(user, src, SKILL_TASK_DIFFICULT, list(/datum/skill/construction = SKILL_LEVEL_TRAINED), -0.2))
-		// to-do: sound
 		ChangeTurf(/turf/simulated/floor/plating)
 		return
 
 	if(isscrewing(C))
-		// to-do: sound
-		if(!user.is_busy() && do_skilled(user, src, SKILL_TASK_DIFFICULT, list(/datum/skill/construction = SKILL_LEVEL_TRAINED), -0.2))
-			qdel(src)
 		// Parent also has screwdriver disassembly so we ought to stop here...
+		to_chat(user, "<span class='warning'>What would that do to a forcefield?</span>")
+		return
+
+	if(ispulsing(C) && !user.is_busy() && do_skilled(user, src, SKILL_TASK_DIFFICULT, list(/datum/skill/construction = SKILL_LEVEL_TRAINED), -0.2))
+		qdel(src)
 		return
 
 	return ..()
@@ -61,6 +65,7 @@
 
 /obj/structure/replicator_forcefield/Destroy()
 	// to-do: sound
+	playsound(src, pick('sound/machines/arcade/gethit1.ogg', 'sound/machines/arcade/gethit2.ogg', 'sound/machines/arcade/-mana1.ogg', 'sound/machines/arcade/-mana2.ogg'), VOL_EFFECTS_MASTER, 80)
 	if(!(locate(/obj/structure/stabilization_field) in loc))
 		new /obj/structure/stabilization_field(loc)
 	return ..()
@@ -94,8 +99,9 @@
 	resistance_flags = FULL_INDESTRUCTIBLE
 
 /obj/structure/stabilization_field/attackby(obj/item/C, mob/user)
-	if(isscrewing(C) && !user.is_busy() && do_skilled(user, src, SKILL_TASK_DIFFICULT, list(/datum/skill/construction = SKILL_LEVEL_TRAINED), -0.2))
+	if(ispulsing(C) && !user.is_busy() && do_skilled(user, src, SKILL_TASK_DIFFICULT, list(/datum/skill/construction = SKILL_LEVEL_TRAINED), -0.2))
 		// to-do: sound
+		playsound(src, pick('sound/machines/arcade/gethit1.ogg', 'sound/machines/arcade/gethit2.ogg', 'sound/machines/arcade/-mana1.ogg', 'sound/machines/arcade/-mana2.ogg'), VOL_EFFECTS_MASTER, 80)
 		qdel(src)
 		return
 
@@ -120,6 +126,8 @@
 	resistance_flags = CAN_BE_HIT | FIRE_PROOF
 
 /obj/structure/replicator_barricade/Destroy()
+	// to-do: sound
+	playsound(src, pick('sound/machines/arcade/gethit1.ogg', 'sound/machines/arcade/gethit2.ogg', 'sound/machines/arcade/-mana1.ogg', 'sound/machines/arcade/-mana2.ogg'), VOL_EFFECTS_MASTER, 80)
 	if(!(locate(/obj/structure/stabilization_field) in loc))
 		new /obj/structure/stabilization_field(loc)
 	return ..()
@@ -161,6 +169,8 @@ ADD_TO_GLOBAL_LIST(/obj/structure/forcefield_node, forcefield_nodes)
 		global.area2free_forcefield_nodes[A.name] += 1
 
 /obj/structure/forcefield_node/Destroy()
+	// to-do: sound
+	playsound(src, 'sound/machines/arcade/heal2.ogg', VOL_EFFECTS_MASTER, 80)
 	global.replicators_faction.nodes_to_spawn += 1
 
 	var/area/A = get_area(src)
